requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-spectacular

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
django-environ==0.11.2

# Caching
redis==4.5.5
django-redis==5.3.0

# Celery
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Filtering
django-filter==23.2

# # Development (optional)
# pytest==7.4.0
# pytest-django==4.7.0
# factory-boy==3.2.1